Switch from custom result type to Swift 5's built-in `Result`.
